feat: rename Reports => Users
